+++ samples/amqscb.go
@@ -1,14 +1,12 @@
 /*
- * This is an example of a Go program to put messages to an IBM MQ
- * queue.
+ * This is an example of a Go program to get messages from an IBM MQ
+ * queue. It uses the asynchronous callback operation instead of using
+ * a synchronous MQGET.
  *
  * The queue and queue manager name can be given as parameters on the
  * command line. Defaults are coded in the program.
  *
- * A single message is put, containing a "hello" and timestamp.
- * Each MQI call prints its success or failure. The MsgId of the
- * put message is also printed so it can be used as an optional
- * selection criterion on the amqsget sample program.
+ * Each MQI call prints its success or failure.
  *
  */
 package main
@@ -33,7 +31,6 @@ package main
 */
 
 import (
-	"encoding/hex"
 	"fmt"
 	"os"
 	"strings"
@@ -45,9 +42,25 @@ import (
 var qMgrObject ibmmq.MQObject
 var qObject ibmmq.MQObject
 
-// Main function that simply calls a subfunction to ensure defer routines are called before os.Exit happens
+var ok = true
+
+// The main function just expects to be given a return code for Exit()
 func main() {
 	os.Exit(mainWithRc())
+}
+
+// This is the callback function invoked when a message arrives on the queue.
+func cb(hObj *ibmmq.MQObject, md *ibmmq.MQMD, gmo *ibmmq.MQGMO, buffer []byte, cbc *ibmmq.MQCBC, err *ibmmq.MQReturn) {
+	buflen := len(buffer)
+	if err.MQCC != ibmmq.MQCC_OK {
+		fmt.Println(err)
+		ok = false
+	} else {
+		// Assume the message is a printable string, which it will be
+		// if it's been created by the amqsput program
+		fmt.Printf("In callback - Got message of length %d from queue %s: ", buflen, hObj.Name)
+		fmt.Println(strings.TrimSpace(string(buffer[:buflen])))
+	}
 }
 
 // The real main function is here to set a return code.
@@ -57,7 +70,7 @@ func mainWithRc() int {
 	qMgrName := "QM1"
 	qName := "DEV.QUEUE.1"
 
-	fmt.Println("Sample AMQSPUT.GO start")
+	fmt.Println("Sample AMQSCB.GO start")
 
 	// Get the queue and queue manager names from command line for overriding
 	// the defaults. Parameters are not required.
@@ -69,15 +82,11 @@ func mainWithRc() int {
 		qMgrName = os.Args[2]
 	}
 
-	// This is where we connect to the queue manager. It is assumed
-	// that the queue manager is either local, or you have set the
-	// client connection information externally eg via a CCDT or the
-	// MQSERVER environment variable
+	// Connect to the queue manager.
 	qMgrObject, err := ibmmq.Conn(qMgrName)
 	if err != nil {
 		fmt.Println(err)
 	} else {
-		// Make sure we disconnect from the queue manager later
 		fmt.Printf("Connected to queue manager %s\n", qMgrName)
 		defer disc(qMgrObject)
 	}
@@ -87,9 +96,9 @@ func mainWithRc() int {
 		// Create the Object Descriptor that allows us to give the queue name
 		mqod := ibmmq.NewMQOD()
 
-		// We have to say how we are going to use this queue. In this case, to PUT
+		// We have to say how we are going to use this queue. In this case, to GET
 		// messages. That is done in the openOptions parameter.
-		openOptions := ibmmq.MQOO_OUTPUT
+		openOptions := ibmmq.MQOO_INPUT_EXCLUSIVE
 
 		// Opening a QUEUE (rather than a Topic or other object type) and give the name
 		mqod.ObjectType = ibmmq.MQOT_Q
@@ -99,47 +108,57 @@ func mainWithRc() int {
 		if err != nil {
 			fmt.Println(err)
 		} else {
-			// Make sure we close the queue once we're done with it
 			fmt.Println("Opened queue", qObject.Name)
 			defer close(qObject)
 		}
 	}
 
-	// PUT a message to the queue
 	if err == nil {
-		// The PUT requires control structures, the Message Descriptor (MQMD)
-		// and Put Options (MQPMO). Create those with default values.
-		putmqmd := ibmmq.NewMQMD()
-		pmo := ibmmq.NewMQPMO()
+		// The GET/MQCB requires control structures, the Message Descriptor (MQMD)
+		// and Get Options (MQGMO). Create those with default values.
+		getmqmd := ibmmq.NewMQMD()
+		gmo := ibmmq.NewMQGMO()
 
 		// The default options are OK, but it's always
 		// a good idea to be explicit about transactional boundaries as
 		// not all platforms behave the same way.
-		pmo.Options = ibmmq.MQPMO_NO_SYNCPOINT
+		gmo.Options = ibmmq.MQGMO_NO_SYNCPOINT
 
-		// Tell MQ what the message body format is. In this case, a text string
-		putmqmd.Format = ibmmq.MQFMT_STRING
+		// Set options to wait for a maximum of 3 seconds for any new message to arrive
+		gmo.Options |= ibmmq.MQGMO_WAIT
+		gmo.WaitInterval = 3 * 1000 // The WaitInterval is in milliseconds
 
-		// And create the contents to include a timestamp just to prove when it was created
-		msgData := "Hello from Go at " + time.Now().Format(time.RFC3339)
+		// The MQCBD structure is used to specify the function to be invoked
+		// when a message arrives on a queue
+		cbd := ibmmq.NewMQCBD()
+		cbd.CallbackFunction = cb // The function at the top of this file
 
-		// The message is always sent as bytes, so has to be converted before the PUT.
-		buffer := []byte(msgData)
+		// Register the callback function along with any selection criteria from the
+		// MQMD and MQGMO parameters
+		err = qObject.CB(ibmmq.MQOP_REGISTER, cbd, getmqmd, gmo)
+	}
 
-		// Now put the message to the queue
-		err = qObject.Put(putmqmd, pmo, buffer)
-
-		if err != nil {
-			fmt.Println(err)
-		} else {
-			fmt.Println("Put message to", strings.TrimSpace(qObject.Name))
-			// Print the MsgId so it can be used as a parameter to amqsget
-			fmt.Println("MsgId:" + hex.EncodeToString(putmqmd.MsgId))
+	if err == nil {
+		// Then we are ready to enable the callback function. Any messages
+		// on the queue will be sent to the callback
+		ctlo := ibmmq.NewMQCTLO() // Default parameters are OK
+		err = qMgrObject.Ctl(ibmmq.MQOP_START, ctlo)
+		if err == nil {
+			// Use defer to disable the message consumer when we are ready to exit.
+			// Otherwise the shutdown will give MQRC_HCONN_ASYNC_ACTIVE error
+			defer qMgrObject.Ctl(ibmmq.MQOP_STOP, ctlo)
 		}
 	}
 
+	// Keep the program running until the callback has indicated there are no
+	// more messages.
+	d, _ := time.ParseDuration("5s")
+	for ok && err == nil {
+		time.Sleep(d)
+	}
+
 	// Exit with any return code extracted from the failing MQI call.
-	// Deferred disconnect will happen after the return
+	// Deferred disconnect/close will happen after the return
 	mqret := 0
 	if err != nil {
 		mqret = int((err.(*ibmmq.MQReturn)).MQCC)